[fix] full layout - finally got timezones sorted
Apparently trmnl does timezones based on your account properties
in their servers, so we have to handle it on our end. Fine. The
confusion happened because if you set trmnlp's configuration with
a timezone, it's off. trmnlp serve uses the local system time as
UTC (feels like a bug), so setting a time_zone in the config
makes things wrong. There's 4 hours of my life I'm not getting
back.